recipes-utils: ultra96-ap-setup: fix scripts
- fix a race condition with the new wilc 15.5 firmware: the p2p0
interface was taking more time to setup.So Connman was launched before
the p2p0 interface, to configure the wifi (wlan0 interface). Then, when
the p2p0 interface was ready, it was stopping the wifi connection on
wlan0.

To fix it, the commands that were executed by the 'ap.sh' script are now
directly called in the init.d script (ultra96-ap-setup.sh). So that,
they are now blocking.

- remove the udhcpd patch. Modifications are performed directly on the
conf file (udhcpd.conf)

- in the bbappend: INITSCRIPT_PARAMS is modified to add the init.d
script in runlevel 0 and 6, so that it will be executed when powering
down the board.