recipes-kernel: linux: uz7ev-evcc-quadcam-h: add patch to handle REV1 serializers
The new REV1 serializers have the HIM pin high, so we have to setup the
serializers and deserializer to HIM mode.

The serializers can be set with HIM off or on (Rev0 or Rev1).
HACK: To unify all serializers, we send a broadcast signal from the
max9286 driver to force serializers to HIM ON